requirements.txt
datetime
discord.py
feedparser
openai
matplotlib
newspaper3k
trafilatura
lxml[html_clean]
